moar test cleanup
