--- gcc/testsuite/g++.dg/contracts/cpp26/P3510-1.C
@@ -1,15 +1,0 @@
-// { dg-do run }
-// { dg-options "-std=c++23 -fcontracts " }
-
-struct S{
-  bool f() const { return true;}
-  bool i = true;
-  bool k = true;
-  S() pre(&(this->k) != &(this->i)) post(i) {};
-  ~S() pre(f()){};
-};
-int main()
-{
-  S s;
-
-}
