+++ gcc/testsuite/g++.dg/contracts/cpp26/basic.contract.eval.p17-SMF4.C
@@ -1,9 +1,16 @@
-// Throwing violation handler in a pre/post check on a noexcept function
-// behaves as if the function exited via an exception.
+// N5008:
+// basic.contract.eval/p17
+// If a contract-violation handler invoked from the evaluation of a function contract assertion (9.4.1) exits via
+// an exception, the behavior is as if the function body exits via that same exception.
+// [Note 12 : A function-try-block (14.1) is the function body when present and thus does not have an opportunity to
+// catch the exception. If the function has a non-throwing exception specification, the function std::terminate is
+// invoked (14.6.2). — end note]
+//
 // This tests the behaviour of a pre condition on an implictly noexcept
 // destructor
-// { dg-do run }
-// { dg-options "-std=c++2a -fcontracts -fcontract-evaluation-semantic=observe " }
+//
+// { dg-do run { target c++23 } }
+// { dg-additional-options "-fcontracts -fcontract-evaluation-semantic=observe " }
 
 #include <contracts>
 #include <exception>
