+++ gcc/testsuite/g++.dg/contracts/cpp26/basic.contract.eval.p6.observe.C
@@ -1,6 +1,6 @@
 // N5008 :
-// basic.contract.eval/p4
-//There is an observable checkpoint ([intro.abstract]) C that happens before
+// basic.contract.eval/p6
+// There is an observable checkpoint ([intro.abstract]) C that happens before
 // (evaluation of a contract assertion ) A such that any other operation O
 // that happens before A also happens before C
 // { dg-do compile { target c++23 } }
