+++ gcc/testsuite/g++.dg/contracts/cpp26/over.call.func.p3.1.C
@@ -1,5 +1,10 @@
-// { dg-do compile }
-// { dg-options "-std=c++23 -fcontracts " }
+// N5008 :
+// [over.call.func/p3.1
+// if the unqualified function call appears in a precondition assertion of a constructor or a postcondition
+// assertion of a destructor and overload resolution selects a non-static member function, the call is
+// ill-formed;
+// { dg-do compile { target c++23 } }
+// { dg-additional-options "-fcontracts" }
 
 struct S{
   bool f();
@@ -17,6 +22,14 @@ struct S{
             SIn sin;
             return sin.f();
   }());
+};
+
+struct S2{
+  bool f() const { return true;}
+  bool i = true;
+  bool k = true;
+  S2() pre(&(this->k) != &(this->i)) post(i) {};
+  ~S2() pre(f()){};
 };
 int main()
 {
