--- gcc/testsuite/g++.dg/contracts/cpp26/virtual_func_none.C
@@ -1,16 +1,0 @@
-// { dg-do compile }
-// { dg-options "-std=c++2b -fcontracts " }
-
-struct Base
-{
-  virtual int f1() pre(true); // { dg-error "contracts cannot be added to virtual functions" }
-  virtual int f2();
-
-};
-struct Child : Base
-{
-  int f2() pre(true); // { dg-error "contracts cannot be added to virtual functions" }
-
-};
-
-
